LIBITD-2455. Renamed "core/handling.py" to "core/gateway.py"
Renamed "core/handling.py" to more accurate "core/gateway.py", now
that "TopTextbooksProcessor" has been moved into its own file.

https://umd-dit.atlassian.net/browse/LIBITD-2455
